# Optimize PushDownFilter to avoid recreating schema columns